feat(licence): add microsoft vscode
Visual Studio Code is a code editor redefined
and optimized for building and debugging modern
web and cloud applications.
In this commit, the Microsoft VSCode editor has
been added to the license file in the Gentoo installation.
This commit adds support for using Microsoft VSCode editor
and ensures that it is properly licensed.